Development - 1.0.1 (#14)
* Add demo

* Disable anonymous actions

* Update README.md

* Bad key/value config

* Bad permission handling

* Correcting invalid permissions check; Correcting document cleaner

---------

+++ src/socket_manager.py
@@ -2,9 +2,13 @@ import base64
 import binascii
 import os
 import shutil
+import time
+
 from flask import session, request
 from flask_socketio import SocketIO, emit, leave_room, rooms
 from flask_socketio.namespace import Namespace
+from objects.db_document import DbDocument
+
 from converter import Converter
 from auth_manager import AuthManager, User
 from db_manager import DbManager
@@ -14,6 +18,7 @@ from usefull import gen_random_filename
 import logging
 from rooms_manager import RoomsManager
 from documents_manager import DocumentManager
+from objects.app_config import AppConfig
 
 
 
@@ -24,7 +29,9 @@ class SocketManager (Namespace):
     db_manager: DbManager
     converter: Converter
 
-    def __init__(self, socketio: SocketIO, auth_manager: AuthManager, db_manager: DbManager, converter: Converter, document_manager: DocumentManager, rooms_manager: RoomsManager, data_folder: str):
+    def __init__(self, socketio: SocketIO, auth_manager: AuthManager, db_manager: DbManager, converter: Converter,
+                 document_manager: DocumentManager, rooms_manager: RoomsManager, data_folder: str,
+                 app_config: AppConfig):
         self.socketio = socketio
         self.auth_manager = auth_manager
         self.db_manager = db_manager
@@ -33,6 +40,7 @@ class SocketManager (Namespace):
         self.rooms_manager = rooms_manager
         self.logger = logging.getLogger()
         self.data_folder = data_folder
+        self.app_config = app_config
         super().__init__()
 
     def on_connect(self):
@@ -45,22 +53,32 @@ class SocketManager (Namespace):
 
     ## Document Editor
     def on_join_room(self, doc_uuid):
-        u_info = self.auth_manager.get_userinfos()
+        def _():
+            u_info = self.auth_manager.get_userinfos()
+            u_uid = self.auth_manager.get_current_u_uid()
 
-        if self.db_manager.get_document(doc_uuid, u_info.user_unique_identifier) is not None:
-            emit("join_room", "OK")
+            #if not self.app_config.allow_anonymous_edit and u_uid == 0:
+            #    emit("display_error", {"error": "Anonymous edition is disabled"})
+            #    emit("join_room", "KO")
+            #    self.logger.error("Anonymous user trying to edit, refusing")
+            #    return
+            print(doc_uuid)
 
-            self.logger.info(f"User '{u_info.username}' join room '{doc_uuid}'")
+            if self.db_manager.get_document(doc_uuid, u_uid) is not None:
+                emit("join_room", "OK")
 
-            self.rooms_manager.add_user_to_room(doc_uuid, u_info)
+                self.logger.info(f"User '{u_info.username}' join room '{doc_uuid}'")
 
-            self.socketio.emit("room_info", f"{u_info.username} has entered the room", to=doc_uuid)
-            room_users_json = self.rooms_manager.get_room_users(doc_uuid, to_json=True)
+                self.rooms_manager.add_user_to_room(doc_uuid, u_info)
 
-            self.socketio.emit("room_users", room_users_json)
-        else:
-            emit("join_room", "KO")
-            emit("display_error", {"error": "This room doesn't exists or you are not authorized to consult it", "required_action": "reload"})
+                self.socketio.emit("room_info", f"{u_info.username} has entered the room", to=doc_uuid)
+                room_users_json = self.rooms_manager.get_room_users(doc_uuid, to_json=True)
+
+                self.socketio.emit("room_users", room_users_json)
+            else:
+                emit("join_room", "KO")
+                emit("display_error", {"error": "This room doesn't exists or you are not authorized to consult it", "required_action": "reload"})
+        return _()
 
     def on_leave_room(self, doc_uuid):
         @self.rooms_manager.require_editing_room
@@ -114,11 +132,7 @@ class SocketManager (Namespace):
         @self.rooms_manager.require_write_permission
         def _():
             doc_uuid = self.document_manager.get_user_current_doc_uuid(rooms())
-            u_info = self.auth_manager.get_userinfos()
-            if u_info is not None:
-                u_uid = u_info.user_unique_identifier
-            else:
-                u_uid = 0
+            u_uid = self.auth_manager.get_current_u_uid()
             if not self.db_manager.get_document(doc_uuid, u_uid):
                 self.logger.info("User tried to push image with invalid doc uuid or u_uid")
                 emit("Unable to retrieve the document with your permissions")
@@ -151,20 +165,12 @@ class SocketManager (Namespace):
         return _()
 
     def on_set_document_name(self, data):
+        @self.rooms_manager.require_write_permission
         def _():
-            doc_uuid = self.document_manager.get_user_current_doc_uuid(rooms())
-            if doc_uuid is None:
-                doc_uuid = data["document"]["doc_uuid"]
-            u_info = self.auth_manager.get_userinfos()
-            if u_info is not None:
-                u_uid = u_info.user_unique_identifier
-            else:
-                u_uid = 0
             new_name = data["document"]["doc_name"]
-            document = self.db_manager.get_document(doc_uuid, u_uid)
+            document = self.user_get_document(doc_uuid, write_required=True)
             if document is None:
                 self.logger.error(f"Error setting document name : Document '{doc_uuid}' not found in DB for uuid '{u_uid}'")
-                emit("display_error", {"error": "This document was not found for your uuid"})
                 return
             self.logger.info(f"Changing document {doc_uuid} name : {document.doc_name} -> {new_name}")
             document.doc_name = new_name
@@ -172,23 +178,23 @@ class SocketManager (Namespace):
                 emit("display_error", {"error": "You are not able to set name for this document"})
                 return
             self.db_manager.set_document(document)
+
+        doc_uuid = self.document_manager.get_user_current_doc_uuid(rooms())
+        if doc_uuid is None:
+            doc_uuid = data["document"]["doc_uuid"]
+        u_uid = self.auth_manager.get_current_u_uid()
+        try:
+            return _()
+        finally:
             new_doc = self.db_manager.get_document(doc_uuid, u_uid).json()
             self.socketio.emit("update_document", {"document": new_doc}, to=doc_uuid)
             emit("set_document_name", {"document": new_doc})
-        return _()
 
     def on_set_document_restriction(self, data):
         @self.rooms_manager.require_editing_room
         @self.rooms_manager.require_write_permission
         def _():
-            doc_uuid = self.document_manager.get_user_current_doc_uuid(rooms())
-            u_info = self.auth_manager.get_userinfos()
-            if u_info is not None:
-                u_uid = u_info.user_unique_identifier
-            else:
-                u_uid = 0
             new_restriction = data["document"]["restriction"]
-            document = self.db_manager.get_document(doc_uuid, u_uid)
             if document is None:
                 self.logger.error(
                     f"Error setting document restriction : Document '{doc_uuid}' not found in DB for uuid '{u_uid}'")
@@ -200,8 +206,15 @@ class SocketManager (Namespace):
                 self.logger.info(f"Changing document {doc_uuid} restriction : {document.restriction} -> {new_restriction}")
                 document.restriction = new_restriction
                 self.db_manager.set_document(document)
-            self.socketio.emit("update_document", {"document": self.db_manager.get_document(doc_uuid, u_uid).json()}, to=doc_uuid)
-        return _()
+
+        doc_uuid = self.document_manager.get_user_current_doc_uuid(rooms())
+        u_uid = self.auth_manager.get_current_u_uid()
+        document = self.db_manager.get_document(doc_uuid, u_uid)
+        try:
+            return _()
+        finally:
+            self.socketio.emit("update_document", {"document": self.db_manager.get_document(doc_uuid, u_uid).json()},
+                               to=doc_uuid)
 
     def on_download_document(self, data):
         @self.rooms_manager.require_editing_room
@@ -229,11 +242,7 @@ class SocketManager (Namespace):
         @self.rooms_manager.require_editing_room
         def _():
             doc_uuid = self.document_manager.get_user_current_doc_uuid(rooms())
-            u_info = self.auth_manager.get_userinfos()
-            if u_info is not None:
-                u_uid = u_info.user_unique_identifier
-            else:
-                u_uid = 0
+            u_uid = self.auth_manager.get_current_u_uid()
             doc = self.db_manager.get_document(doc_uuid, u_uid)
             with open(os.path.join(self.document_manager.get_document_folder(doc_uuid), doc_uuid + ".adoc"), 'r') as f:
                 content = f.read()
@@ -251,6 +260,9 @@ class SocketManager (Namespace):
             }
             emit("init_document_editor", ret)
         return _()
+
+    def on_get_app_config(self):
+        emit("get_app_config", self.app_config.json())
 
     ## Authent
     def on_get_auth_methods(self):
@@ -275,35 +287,32 @@ class SocketManager (Namespace):
         if isinstance(data, dict):
             if data.get("example"):
                 example = True
-        u_info = self.auth_manager.get_userinfos()
-        if u_info is not None:
-            u_uid = u_info.user_unique_identifier
-        else:
-            u_uid = 0
+        u_uid = self.auth_manager.get_current_u_uid()
 
         if not example:
-            conf_permission = os.environ.get("WASCII_DEFAULT_DOC_PERMISSION")
-            if conf_permission is None:
-                initial_restriction = DocumentRestriction.EDITABLE
-            else:
-                initial_restriction = DocumentRestriction.from_string(conf_permission)
-                if initial_restriction is None:
-                    self.logger.warning("Invalid document permission set in config, defaulting to EDITABLE")
+            if (not self.app_config.allow_anonymous_creation or not self.app_config.allow_anonymous_edit) and u_uid == 0:
+                emit("display_error", {"error": "Anonymous creation is not allowed"})
+                return
+            initial_restriction = self.app_config.default_doc_permission
+            print(initial_restriction)
 
             new_doc = self.db_manager.create_document("New document", u_uid, initial_restriction)
             self.logger.info(f"Created document : {new_doc.json()}")
             os.mkdir(self.document_manager.get_document_folder(new_doc.doc_uuid))
             # with open(os.path.join(tmp_folder, new_doc.doc_uuid + ".adoc"), 'w') as f:  # Create the file
-            with open(os.path.join(self.document_manager.get_document_folder(new_doc.doc_uuid), new_doc.doc_uuid + ".adoc"),
+            with open(os.path.join(self.document_manager.get_document_folder(new_doc.doc_uuid),
+                                   new_doc.doc_uuid + ".adoc"),
                       'w') as f:  # Create the file
                 f.write("")
         else:
-            new_doc = self.db_manager.create_document("Example document", u_uid, DocumentRestriction.PRIVATE)
+            new_doc = self.db_manager.create_document("Example document", u_uid,
+                                                      DocumentRestriction.PRIVATE, temporary=True)
             self.logger.info(f"Created example document : {new_doc.json()}")
             with open(Path(__file__).parent.joinpath("example_document.adoc"), 'r') as f:
                 doc = f.read()
             os.mkdir(self.document_manager.get_document_folder(new_doc.doc_uuid))
-            with open(os.path.join(self.document_manager.get_document_folder(new_doc.doc_uuid), new_doc.doc_uuid + ".adoc"),
+            with open(os.path.join(self.document_manager.get_document_folder(new_doc.doc_uuid),
+                                   new_doc.doc_uuid + ".adoc"),
                       'w') as f:  # Create the file
                 f.write(doc)
         emit("create_document", {"document": new_doc.json()})
@@ -311,17 +320,14 @@ class SocketManager (Namespace):
     def on_delete_document(self, data):
         try:
             doc_uuid = data.get("document").get("doc_uuid")
-            u_info = self.auth_manager.get_userinfos()
-            if u_info is not None:
-                u_uid = u_info.user_unique_identifier
-            else:
-                u_uid = 0
+            u_uid = self.auth_manager.get_current_u_uid()
+
+            if not self.app_config.allow_anonymous_edit and u_uid == 0:
+                emit("display_error", {"error": "Anonymous edit is not allowed"})
+                return
 
             doc = self.db_manager.get_document(user_uuid=u_uid, doc_uuid=doc_uuid)
             if doc is None:
-                self.socketio.emit("display_error",
-                              {"error": "Document does not exist or you are not authorized to delete it"})
-                self.logger.warning(f"Document {doc_uuid} not found for user {u_uid}")
                 return
             doc_name = doc.doc_name
             if not self.db_manager.delete_document(u_uid, doc_uuid):
@@ -337,11 +343,24 @@ class SocketManager (Namespace):
 
     def on_list_documents(self):
         # print(sid)
-        u_info = self.auth_manager.get_userinfos()
-        if u_info is not None:
-            u_uid = u_info.user_unique_identifier
-        else:
-            u_uid = 0
+        u_uid = self.auth_manager.get_current_u_uid()
         documents = self.db_manager.list_documents(u_uid)
         ret = [doc.json() for doc in documents]
         emit("list_documents", ret)
+
+    def user_get_document(self, doc_uuid, write_required=False) -> DbDocument | None:
+        u_uid = self.auth_manager.get_current_u_uid()
+
+        doc = self.db_manager.get_document(user_uuid=u_uid, doc_uuid=doc_uuid)
+        if doc is None:
+            self.socketio.emit("display_error",
+                               {"error": "Document does not exist or you are not authorized to access it"})
+            self.logger.warning(f"Document {doc_uuid} not found for user {u_uid}")
+            return
+        if write_required:
+            if not self.document_manager.can_user_write_document(doc_uuid, u_uid):
+                self.logger.warning(f"Document {doc_uuid} can't be written by {u_uid}")
+                emit("display_error", {"error": "You are not authorized to write this document"})
+                return
+        return doc
+
